fix : crud product and service

+++ src/lib/action/product.ts
@@ -28,7 +28,7 @@ export const createProduct = async (data: ProductSchema, image?: FormData) => {
 
     if (image) {
       const result = await uploadImage(image);
-      if (result.error) {
+      if (result.error.status) {
         return responServerAction({
           statusError: true,
           messageError: result.error.message,
@@ -94,31 +94,33 @@ export const updateProduct = async (
       });
     }
 
-    let imagePath = oldProduct.image; // Default tetap pakai gambar lama
+    let imagePath = oldProduct.image;
 
-    // Jika ada image baru yang diunggah
     if (image) {
-      const result = await uploadImage(image);
-      if (result.error.status) {
-        return responServerAction({
-          statusError: true,
-          messageError: result.error.message,
-        });
-      }
-
-      // Hapus gambar lama jika ada
-      if (oldProduct.image) {
-        const oldImagePath = path.join(
-          process.cwd(),
-          "public/uploads",
-          oldProduct.image
-        );
-        if (fs.existsSync(oldImagePath)) {
-          fs.unlinkSync(oldImagePath);
+      const imageFile = image.get("image");
+      if (imageFile && imageFile instanceof File) {
+        const result = await uploadImage(image);
+        if (result.error.status) {
+          return responServerAction({
+            statusError: true,
+            messageError: result.error.message,
+          });
         }
-      }
 
-      imagePath = result.data; // Update dengan image baru
+        // Hapus gambar lama jika ada
+        if (oldProduct.image) {
+          const oldImagePath = path.join(
+            process.cwd(),
+            "public/uploads",
+            oldProduct.image
+          );
+          if (fs.existsSync(oldImagePath)) {
+            fs.unlinkSync(oldImagePath);
+          }
+        }
+
+        imagePath = result.data;
+      }
     }
 
     await prisma.produk.update({
